TBD_CG3/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dacite==1.8.1
distro==1.7.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
Flask==3.0.3
fonttools==4.55.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.20.2
httpx==0.27.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==8.10.0
numpy==2.1.3
oauthlib==3.2.0
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pokebase==1.3.0
pokemontcgsdk==3.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.10
pydantic_core==2.33.2
pygame==2.6.1
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML>=6.0
requests==2.32.3
rsa==4.9.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.42.0
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
uvicorn==0.37.0
Werkzeug==3.1.1
zipp==1.0.0
